Rename team packet package
