+++ src/main/java/kernitus/plugin/OldCombatMechanics/utilities/packet/team/PostV17TeamPacket.java
@@ -1,4 +1,4 @@
-package kernitus.plugin.OldCombatMechanics.utilities.packet.fun;
+package kernitus.plugin.OldCombatMechanics.utilities.packet.team;
 
 import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketHelper;
 import kernitus.plugin.OldCombatMechanics.utilities.reflection.Reflector;
